DAOs Review das Reviews

+++ app/src/main/java/studentcompany/sportgest/daos/Attribute_Exercise_DAO.java
@@ -70,6 +70,8 @@ public class Attribute_Exercise_DAO extends GenericPairDAO<Attribute, Exercise> 
             res.moveToNext();
         }
 
+        res.close();
+
         return resList;
     }
 
@@ -89,15 +91,7 @@ public class Attribute_Exercise_DAO extends GenericPairDAO<Attribute, Exercise> 
             resList.add(exercise_dao.getById(exerciseId));
             res.moveToNext();
         }
-
-        //Sorting
-        Collections.sort(resList, new Comparator<Exercise>() {
-            @Override
-            public int compare(Exercise ex1, Exercise ex2) {
-
-                return ex1.getTitle().compareTo(ex2.getTitle());
-            }
-        });
+        res.close();
 
         return resList;
     }
@@ -119,20 +113,20 @@ public class Attribute_Exercise_DAO extends GenericPairDAO<Attribute, Exercise> 
             res.moveToNext();
         }
 
-        //Sorting
-        Collections.sort(resList, new Comparator<Attribute>() {
-            @Override
-            public int compare(Attribute at1, Attribute at2) {
-
-                return at1.getName().compareTo(at2.getName());
-            }
-        });
+        res.close();
 
         return resList;
     }
 
     @Override
     public long insert(Pair<Attribute, Exercise> object) throws GenericDAOException {
+
+        if(object==null)
+            return -1;
+
+        if(object.getFirst() == null || object.getSecond() == null)
+            return -1;
+
         ContentValues contentValues = new ContentValues();
         contentValues.put(COLUMN_ATTRIBUTE_ID, object.getFirst().getId());
         contentValues.put(COLUMN_EXERCISE_ID, object.getSecond().getId());
@@ -142,6 +136,13 @@ public class Attribute_Exercise_DAO extends GenericPairDAO<Attribute, Exercise> 
 
     @Override
     public boolean delete(Pair<Attribute, Exercise> object) throws GenericDAOException {
+
+        if(object==null)
+            return false;
+
+        if(object.getFirst() == null || object.getSecond() == null)
+            return false;
+
         int deletedCount = db.delete(TABLE_NAME,
                 COLUMN_ATTRIBUTE_ID + " = ? AND " + COLUMN_EXERCISE_ID + " = ? ",
                 new String[] { Long.toString(object.getFirst().getId()), Long.toString(object.getSecond().getId()) });
@@ -150,14 +151,17 @@ public class Attribute_Exercise_DAO extends GenericPairDAO<Attribute, Exercise> 
 
     @Override
     public boolean exists(Pair<Attribute, Exercise> object) throws GenericDAOException {
+
         if(object==null)
             return false;
 
+        if(object.getFirst() == null || object.getSecond() == null)
+            return false;
+
         int fields = 0;
-        int tmpInt;
         long tmpLong;
 
-        StringBuilder statement = new StringBuilder("SELECT * FROM "+ TABLE_NAME +" where ");
+        StringBuilder statement = new StringBuilder("SELECT * FROM "+ TABLE_NAME +" WHERE ");
         if ((tmpLong = object.getFirst().getId()) >= 0) {
             statement.append(COLUMN_ATTRIBUTE_ID + "=" + tmpLong);
             fields++;
